Core Cleanup #6
Added default constructor to Material class and fixed afew other things related to the class. There is tons of work I realise I need to do regarding the structure of the Material and Texture classes as these are currently spaghetti code.

Updated public function for Transform to capitalisation convention for GetTransform. Additionally I cleaned some scenes so that it called the =operator to return a mat4 opposed to calling GetTransform (even though the =operator calls this function anyway) because it makes the scene code look cleaner.

Added TODO for Shader class and Material class.

Will merge Core Cleanup branch to master after these changes as I am extremely happy with the changes so far made to the core functionality of the engine.