Forward+ Rendering #2
Making lots of changes to the engine functionality regarding Scene Management, Materials, Shaders, Rendering, etc.

Implementation of the Scene specific resource manager will be to ensure that the scene can load everything it needs, whilst not being weighed down by the redundancy of resources from other scenes that are not loaded.

Started working on the forward plus rendering pipeline in the scene update function. Light SSBOs are working with the shaders which is good, sorting by materials is working, not yet sure on the Compute Shader side of things, need to read the data that is being created in the light indices per sceen tile.

Struggling with ENTT components in scene with materials and meshes. Struggling with rendering algorithm to reduce overhead of GPU commands for updating uniforms for shaders etc.

The way I want things to work with meshes and materials is that when you load a mesh, it loads its default material into the scenes resources manager that it can refer to. If a mesh then has a change to the material, it then creates a new material within the scenes resources OR checks to see if there are currently any materials with the scene resources manager that are the same and returns a reference to that material in the resources manager.

This then means that I can do draw calls centric to materials present within the scene so that everytime I draw a mesh with the same material, I don't need to keep resetting the uniforms.

WIP still lots to do...