Forward+ Rendering #7
- Core -

Removed old SetUniforms function in Material class and updated all references to the superseded function UpdateUniforms. Changed UpdateUniform to take the Camera class opposed to the CameraComponent struct.

Removed old DrawMeshComponent function from the Renderer class.

Removed OldLight struct from Lights.h.

Expanded declaration of functions within Components.h to a CPP file.

Updated Transform::GetTransform to return a glm::mat4 that is not const, rather the function is const and ensures no private members of Transform are changed. Implemented individual X, Y, and Z functions for Translate, Rotate and Scale.

Rehauled manual MeshRenderer and MeshFilter loading that is not through the scene. Created ManualDraw function to MeshRenderer for manual draw calls of the object. PLEASE NOTE: as this is under the assumption the MeshRenderer is not within a Scene, it will draw using a flat shader with no lighting.

I also want to add Tangent and Bitangent to the VAOs, updated all shaders for this to work and also updated the instancing so that it refers to the correct layout slot in the shader. Currently it is commented out of the Vertex struct as it did not work when I added them. Time to diagnose my code from 4 years ago again!

- Sandbox -

Updated references to Material::SetUniforms to Material::UpdateUniforms(CAMERA).

Fixed Scene 7 so that it MeshRenderer::ManualDraw()'s two models that have been loaded manually and isolated from a Scene.

Added a PointLight with a Torch Model to Scene 8 for better lighting.

Added ManualLight struct for scenes that still use the old lighting convention that has been removed from Core.