Skybox Serialization and Other Changes
- Louron Core -

Cleanup Vertex Array and Vertex Buffer bindings when finished Adding Vertex Buffer to Vertex Array.

Assigned integer values to L_CAMERA_CLEAR_FLAGS enum.

Implemented Serialization and Deserialization for scenes with entities that have SkyboxComponent's and SkyboxMaterial's. This includes Serialization and Deserialization for material files for SkyboxMaterial types.

Moved SerializeEntity into SceneSerializer class for access to m_Scene.

Added clear flag Serialization and Deserialization for CameraComponent.

Updated TextureFilePath vector in SkyboxMaterial class to a fixed array of 6 filesystem::path's for safer handling.

Updated UnBind for SkyboxMaterial.

Fixed VertexArray creation for SkyboxComponent. When SceneSerializer created a new SkyboxComponent, the VertexArray was not created correctly, so I had to put this onto the heap to control the data properly.

Updated SkyboxMaterialin SkyboxComponent to be a shared_ptr as there was issues with the initialisation of the stack allocated SkyboxMaterial.

Updated naming convention of public methods in Camera class. Added UpdateProjMatrix method so that we can update the projection as needed.

Scene class will now UpdateProjMatrix of the primary camera with the most up to date information when the scene is started (I don't want to do this every frame).

- Louron Sandbox -

Added skybox material file for Kloppenheim Skybox.

Scene 12 - Updated lastTime to be reset when scene is attached during runtime (similar to how it's done in Scene 11).

Scene 11 - Updated setup of SkyboxMaterial.

Added UpdateProjMatrix for the cameras in applicable scenes.