Framebuffer Resize Handling
- Louron Core -

Updated Forward+ Render Pipeline to now handle when the framebuffer (currently the GLFW window) is resized.

This updates the screen size, number of tiles, light indices buffer sizes, and the depth texture used all part of the F+ rendering pipeline.

Design has changed so you now need to pass a shared_ptr of the scene instance to RenderPipeline when starting the pipeline. If this is not passed, it will not conduct any rendering.

You can also change this during runtime to another scene if required to change scenes.

RenderPipeline and derived classes now have access to the screen size of the GLFW window (any changes to it) and the scene that is currently set to be rendered.