Entity GetComponent Nullification
- Louron Core -

Updated Entity::GetComponent() method so that if we call this on an entity that does not hold that component, we do not crash the entire game, or halt runtime, we rather pass back a reference to a static blank component of that type.

During runtime, this means that if I have an error in handling entities and components, rather than crashing, I receive a log that advises me of any errors or issues with my scene and entity structures.

Added libraries and include files for spdlog.

- Louron Sandbox -

Updated Scene 8, 11, and 12 that use scenes and entities to not have to check if an entity has a component before using that component.