PhysX Implementation
- Louron Core -

This is somewhat of a rudimentary and very poorly executed implementation of the PhysX library by NVidia.

Engine now has two separate loops. Update which runs every single frame, and FixedUpdate which runs at a specified target of 60hz per second (this is determined by the Time class).

Implemented a Time class to provide overarching time management throughout the engine such as delta time, unscaled delta time, fixed delta time, etc. This also has a time scale so we can effectively speed up time or reduce time all the way to zero which would emulate a pause for everything that takes place in the fixedupdate world.

Played around with the idea of implementing a FBO that the renderer would render to first, then we can decide what to do with the texture texture of the FBO. This requires more work as the Forward+ rendering pipeline has a depth attachment, and my tests did not prove fruitful lol... This will also help greatly with mouse picking :D

Physics Core class is a singleton that is used purely to initialise the physx library. I want to look at using this as the sole singleton of the PhysX library throughout my implementation, however currently, this engine just calls the static reference to the physics library from anywhere required in the code.

Issue with window when minimising fixed, had to atleast return 1 pixel back from the Window getter functions when the size was 0,0 because the window was minimised.

Ammended VertexBuffer constructor when a float pointer to data is passed and not a vector of vertex's.

Texture Class now holds the size of the texture in the class.

Removed auto saving from Project and Scene classes. If you want to save a project, you need to call SaveProject or SaveScene explictly.

Updated Project Init so that you can specify an explicit startup scene.

Implemented cleanup function for instancing data.

If scene does not have a camera, it will make one now automatically.

Created Component base class for all components. This required changes in the renderer for light structs to be expliclty defined within the renderer pipeline to ensure there are no issues with data alignment between CPU and GPU.

Implemented PhysX implementation as follows:
 1. PhysicsWrappers - this wraps the direct API to the library in our engine specific requirements.
 2. Colliders - derived from Components, and hold a reference to the PhysicsWrapper of a PxShape. This controls all aspects of the colliders shape.
 3. Rigidbody - derived from Component, and hold a reference to the PhyicsWrapper of a PxRigidDynamic. This controls all aspects of physics simulation in relation to reacting to physics in the PxScene.

Added Scene Systems for Transform updates and Physics updates per frame. Opposed to updating everything immediately as something is called, we flag the component as dirty using bitwise operations, and we check these flags in a systems update before rendering or processing other scene requirements dependent on these changes.

When components are added to the entity, the entity will check the type of component and perform relevant operations to that particular object for initialisation.

Added serialisation and deserialisation of the physics components.

- Louron Sandbox -

I have created two new scenes.

One that has a basic scene with physics objects that are either green whilst awake, or red whilst sleeping.

The second scene is a remake of my SDD year 12 major work game! This was very fun to relive, and was the main reason I decided to implement the PhysX library, so I could make really rudimentary games that have some physics. I would like to next make a character controller game, and fix how the CameraComponent works currently so that I can define differing behaviours for cameras. E.g., ortho, perspective, FPS cameras, third person cameras, static cameras, cameras that have basic logic like move from A to B, etc.

Fixed alot of the scenes as the changes with the transform component led to issues with getting the udpated transform of the transform.