Custom Collision Callback, Render FBO, Mouse Picking, and Audio Changes
- Louron Core -

Implemented system for custom collision callbacks. This is so that for each shape, you can define a custom callback which is called when that shape collides with something else. Each individual instance of a component collider can have it's own custom callback functionality.

Updated physics logic so that when you apply a force or torque it checks if the scene is currently processing physics simulations. If it is processing physics changes, then we queue the force or torque to be applied next frame. This is for when using multi threaded callbacks that may be called when the scene is simulating physics.

Implemented FBO logic into Forward+ rendering pipeline class. This means that the colour and depth passes are rendered into the FBO opposed to the direct opengl back buffer.

Implemented entity picking texture into FBO. This means that each render pass the entity UUID will be entered into a texture which hold a uint32_t (unsigned int) for each pixel in the screen. We can then query this to gather what entity is at what pixel on the screen, e.g., you pass the x,y for the mouse position when you click, and it returns the UUID of the entity at that location.

Changed UUID from uint64_t to uint32_t as OpenGL may not support 64 bit integers depending on the system. This is to ensure mouse picking works.

Removed IRRKLANG audio library as this was too big for current requirements. We only need something small at the moment which can be statically linked into the Engine library without the need for a shared dll library.

Implemented miniaudio library to replace IRRKLANG.

Updated Camera code so that the ViewMatrix processes whenever there are changes to the camera properties.

Updated ShaderLibrary constructor so that it automatically finds default shader, and updated Engine so that it finds all glsl and comp files within local and recursive directories from the current exectuable path to compile when the engine starts.

Fixed time so that when it initialises, last time is set to now.

Updated Engine entry point so that if it is running a release build, it is a WINAPI WinMain entry point with no console, and vice versa, if debug it will have the console.

- Louron Sandbox -

Updated play audio call in scene 10.

Implemented custom callback in scene 14 so that when two specific entities collide, it applies knockback to the one with least velocity.

Implemented example for mouse picking in scene 14.

Updated UUIDs of Scenes to ensure they are not above uint32_t max!