Hierarchy, Transform, and More Physics
- Louron Core -

So I have implemented a hierarchy system. This will be an independent component in the Entity ECS that will hold information of the parent and a vector of children to the current entity.

This has required a whole lot of changes to how transformations work, as we need to ensure that parent-child transformation relationships are accurate e.g., when the parent transform updates, the child is able to use this to determine its global space relative to the parent and it's own local space coordinates.

Updated alot of the Physics code as Rigidbodies are sort of exempt from being influenced by a parents transformation change. The child with a rigidbody will still have local coordinates relative to the parent, but it's global coordinates will not change based on changes to the parent transform. We also have a mechanism for auto-assignment of colliders to appropriate rigidbodies.

For Example
If I have a collider on Entity B that does not have Rigidbody, it will create it's own internal kinematic rigidbody that does not move unless explictly moved by the transform of the entity.

When Entity B has a parent (Entity A) the physics system will look to auto-assign any child collider components nested within the Hierarchy System to an appropriate parent Rigidbody as another shape and update the inertia and mass of the Rigidbody. This will find the first available Rigidbody in parents. If Entity A did not have a Rigidbody, but it's parent, Entity 1, had a Rigidbody, Entity B's collider would be attached to Entity 1's rigidbody.

Some issues still as a WIP, but it is working alas...
  - Physics RigidActors no longer go to sleep, I think this is just minor rounding issues that lead to the system continuously setting the global or local pose of shapes, and actors.
  - Scale and shapes are on the fritz. If you increase the scale of a parent, the child spherecollider will go absolutely mental... hence on the fritz

(O you don't have the right, o you don't have the right, by the way you don't have the right, o you don't have the right.)

- Louron Sandbox -

Created scene to test and demonstrate hierarchy systems - Scene 15.

All other scenes pretty much just updated Transform API per the change to Louron Core.