PBR Materials and Shading
- Louron Core -

Implemented a ROUGH approach to PBR.

PBR Material subclass has been implemented. Getter, setter, virtual overrides, the lot!

Texture class has slightly been updated so there is a 1x1 pixel blank white texture as the "Default_Texture", and a 1x1 pixel blank normal texture as the "Default_Normal_Texture". TextureLibrary will set these two up on initialisation.

Fixed an issue where eash MeshRenderer held a shared_ptr to a map of materials. This is not what we want because we need different MeshRenderers to have their own list of materials, without impacting other MeshRenderers. When a MeshRenderer is created, it creates a copy of the map of materials. Please note the materials in the map are still a shared_ptr, so if we change that material elsewhere, it will change for our material in meshrenderer::Materials - this is desired behaviour (for now)...

- Louron Sandbox -

I have created a new test scene to demonstrate the PBR materials. This is similar to the LearnOpenGL example with the array of red spheres with differing values, and a textured rusted iron sphere with a PBR material. This scene has mouse picking to update entity transforms, and pressing F wiil toggle a spot light on the camera pointing in the camera's direction.