Asset Management and Prefabs
- Louron Core -

Asset Importer - implemented with afew example basic import methods for different types of assets.

AssetManager - AssetManagerBase and EditorAssetManager classes implemenented. For now these will be the primary classes that are used within Editor based asset management. I do not even want to think about tackling binary serialisation of assets yet smh...

Updating Asset Derived Class - I have ensured that necessary classes throughout the engine have been updated to reflect their are a type of asset, e.g., texture, materials, meshes, etc.

New Mesh System (WIP) - I have implemented a new framework for how I want assets to work. When we import a model file, all meshes are created into their own entities within a prefab object. A single mesh is used per MeshFilterComponent. A single mesh may be made up of multiple sub-meshes that have their own VAO, VBO and EBO. The material that will be utilised for a sub-mesh will depend on the order in which the sub-mesh is in the mesh vector. E.g., 0 of the vector will refer to 0 in the MeshRenderer vector of materials, 1 in the vector will refer to 1 in the MeshRenderer vector of materials, etc. If there are more submeshes than materials, it will use the last iterator of the MeshRenderer. If there are more materials, excess materials will not be applied to the rendering of the sub-meshes. The MeshFilterComponent is not considered an asset, but an ECS component that links to an Asset that may be used for rendering a mesh.

Project Class - Cleaned up the Project class. When we use a Scene, it needs to be guaranteed that there is a Project that is running. The Scene will call static methods from the Project class, so we need to ensure that it is initialised before the scene is.

Renderer Instancing - There were issues with instancing that I discovered when implementing the prefab system. I didn't realise that you cannot use glBufferSubData when you are increasing the size of the buffer, so if there were more transforms introduced, then we would need to reallocate the memory.

Big thanks to Cherno for his design methodology around Asset Management, I have taken alot of his code examples and insight into consideration and will be developing this further in the Louron style.

- Louron Sandbox -

Relocated "assets" folder inside the "Sandbox Project/Assets" folder.

Introduced some more model assets for more variety hehe. Can't remember source of these assets but they've been on my harddrive since like 2016...

Created Scene 17 which is for asset management testing and prefab fun! In this scene you can instantiate torch's that bob up and down by left clicking. This also loads an example model type asset into the new asset manager system so that I can test out importing models from file.