Octree Spatial Partitioning and Other Bits and Pieces
- OctreeEEEEEE -

This probably deserves its own section tbh.

This Octree is completely templated, meaning that I can determine what data source I want to store in the octree, then go from there. This is really good because if I want to use it for Entities on the game engine side for scene partitioning, it can! If I wanted to use it for terrain LOD in the future, I CAN! It was honestly such a headache to implement though due to the recursive nature of an octree.

So basically this is how it works. The OctreeBounds holds a pointer to the root node as a OctreeBoundsNode class.

The root node holds a size_t for the size of data sources contained in this node, the size of data sources contained in this and all child nodes, and a index for where this data begins.

All data is stored within a single global vector in the OctreeBounds class, so validation of node start index and size needs to be updated every single time a change is made to the octree such as inserting or removing data sources.

The Octree will shrink and grow dynamically, so if a data source is attempted to be inserted outside of the octree, it will grow enough to fit that new data source.

If the root node has no data sources and has only one child with data sources, it will shrink.

That is pretty much how it works, but here is a GameDev Stack Exchange question and answer I made in relation to this particular design. Sadly no one answered or commented but myself, but it's a good resource to have up there for years to come if anyone else tries this.

https://gamedev.stackexchange.com/questions/211647/octree-query-frustum-search-and-recursive-vector-inserts

- Louron Core -

Updated Profiler class to include accumulative timers. These will accumulate multiple times throughout a frame, and will be cleared at the start of each frame. This is useful for recursive functions that you want to measure where each profiler used to just overwrite the previous one, and you never see the accumulated time.

Renderer API now has instanced debug cube draw function. We merely pass a vector of glm::mat4's and this function does the rest.

Shadows WIP - nothing serious yet, just an idea of how I want to implement this. Each light will have a shadow type option, and each mesh renderer will have a shadow cast boolean option for if it does or doesn't cast shadows from lights with shadows enabled.

Abstracted bounding aabb, spheres and frustums into their own respective header and source files for easier management.

Forward Plus Pipeline - general cleanup of bits and pieces, made the allocation system for renderable vectors better so that the capacity doubles if it ever comes across a scenario where it needs to reallocate, saving on memory reallocation time by just doing it once and leaving plenty of room to spare. Need to add a cap of how much we can increase the capacity by though, if we keep multiplying by two we could break our computer haha...

Also implemented light frustum culling so that we are only sending the lights we can see to the GPU. Need to add the lights into the Octree somehow. Maybe a separate Octree for purely just lighting? Would this be beneficial or just chuck it in with everything else?

Added glFlush to the LightCull method so that it starts executing the light cull compute shader as soon as possible. I tried double buffering the data to the GPU for this so that it was always working on last frames data, but this did not yield any meaningful results...

Light cull compute shader now uses multiple workers within a workgroup to fill in the final SSBO for each workgroup.

Added ResetMousePosToCurrentPos to camera class. This is so that when we click into the viewport, the camera doesn't keep fritzing out. Camera serialisation now includes clear colour.

Fixed Transform::OnTransformUpdated so that now we actually check if all children MeshFilter AABB's need to be updated, opposed to only marking the parent entity as needing updates to the mesh AABB.

Entity class now has an equality operator to see if the current entity is the same as the other.

Scene class can now check if an entity is within the scene by just passing the Entity opposed to passing the name or UUID.

Added light complexity heatmap toggle so that you can see how complex a certain tile is regarding it's lighting.

- Louron Editor -

Created a few new test scenes, octree tests and light tests. (the main reason I forced myself to implement the octree was because of how poor the performance was when I had 1024 lights in the scene.. I am yet to add lights into the octree, but thought I might make the commit now since I am at a happy state with it).

When viewport is focused we reset the mouse pos in the camera class to ensure it doesn't fritz.

Fixed transform component GUI so that we can actually drag the value now.

Implemented drag drop targets for meshes, materials, and textures in the properties panel.