Minor Changes
- Louron Core -

Created static init and shutdown methods for the audio singleton so that it can be initialised as the engine starts up.

Moved most of the time functions into a CPP source file for faster compilation.

Included OctreeBounds rebuild overload in API with config parameter so that you can rebuild the tree with a passed config set opposed to changing config then rebuilding.