C# Scripting, Physics Refactoring, Editor Changes
- Louron Core -

Implemented mono runtime embedding to enable C# scripting inspired by Cherno's design. This includes the script manager, script connector (to the C# callbacks), script instances, fields, serialisation, etc. Now you can literally make a game in this (WITH LIMTED FUNCTIONALITY lol)

Added functionality to Scene play and stop. When OnRuntimeStart or OnSimulationStart, it will now initialise the appropriate things such as physics scene or script instances that begin functionality of the scene.

Refactored Physics so that it has the correct move, and copy assignment operators and constructors. This was an issue because in copy I was actually moving. I'm so dumb... But now with scene runtime and simulating, it makes it easier because all physics objects are initialised with the data of the component when you want to play the scene, so its not constantly running! Also made adding and removing physics components from entities much easier. Removed physics callbacks per component. This is because we now have scripting, and the callback function that overrides PhysX callback class is now setup to call functions in scripts attached to that entity such as OnCollideEnter, OnTriggerEnter, etc. I should probably make it so it passes data regarding the Collision to the functions in C# because right now all you can really do in C# when you have a collision is get the other object that collided with you, ideally you want data on the contact point, and other things, etc.

Added Copy function to scene so that you can easily copy a Scene. This is useful for in the editor when you want to keep a stored state of the scene you are editing for when you stop running or simulating, just copy the old one back over into the project when stopping the scene!

Fixed some issues with Octree initialisation when creating a new scene, adding new obejcts, and loading scenes.

Fixed issues with Skybox Texture loading - internal formats all need to be the same for all faces of the cubemap, and I was setting the internal format to the format of the texture, so if you loaded a texture that was a different format, the cube map texture wouldn't work and would just be BLACK.

Created some further functionality for getting the local and global directions of a transform, front, up, right.

Refactored alot of the Project system for creating, loading, and saving projects.

Added platform detection - maybe I will use this never but its good to have if I want to start thinking about cross platform design.

Added key codes enum, need to integrate this into the input system.

Updated file utilities for better file saving and opening, and also created new function for selecting only directories.

When the editor asset manager is created, it will automatically load a default Sphere and Cube into memory for insantiation in the scene.

- Louron Editor -

Separated Panels from the main editor layer logic. File was getting too big...

Added PopupModals for alot of things like creating new scenes, projects, selecting entities or entities with certain components for script fields, etc.

Implemented scene control so that you can play and stop a scene in the editor.

Added windows for scene and project settings, e.g., selecting the startup scene for the project.

Added code/file generation helpers for scripting, e.g., generating a new solution project, generating a new C# script file, compiling the app assembly. My thinking is, the engine doesn't need to worry about this shit, only the editor will care because on a deployed game, you're not going to be worrying about generating solutions, or C# files or compiling the code base again, everything you'll ever need will be in binary serialisation.

- Louron Script Core -

Started creating an API for accessing the engine. This includes some standard math things such as Vector2,3,4's math helper functions, callbacks to the engine for components and entities.

Very rudimental at this point in time, but I'm quite proud of what I've managed to do this past week!

Alot more to do here for scripting API functionality, but for now I can probably remake my first game ever with this lol... Maybe for a game jam HAHA!