Content Browser GUI, Script Workspace Changes
- Louron Core -

Added some static methods to AssetManager for utility.

Updated Project to change folder directory generation. Scripts will no longer be part of Assets, but in its own folder relative to Project Directory. I think it's clean to keep a separation between Assets, Scenes, and Scripts.

- Louron Editor -

Implemented Content Browser of my own design! I'm happy with how it is at the moment, but I will probably look to implement a Tree structure for directories and files internally in C++ so that I don't need to have such large overhead for the filesystem. Maybe I only want to check for file/directory changes every 1 or 5 seconds, but every frame! Jeez, that uses alot of processing power...

Content browser is limited to only the Asset, Scene, and Script folders in the directory so you don't have access to the entire PC!

Nice layout with a directory side bar, and a folder/file view being the main centre point on the right. You can drag and drop files into folders, and folders into folders in both the main view and side bar which is awesome.

Another awesome thing is that I started to integrate drag drop functionality with other panels, e.g., you can drag a texture into a SkyboxMaterial, and it will update the texture it uses on that skybox face, or you can drag an FBX file from the content browser into the hierarchy panel, and it will instantiate it in the scene! Awesome stuff honestly... MUAHAH!

It's so relieving to be able to have a content browser honestly because wanting to create a new object with a certain mesh etc., I would have to open the asset registry window and trawl through assets to try and find the right one and make sure its the right type! then create an enttiy, add components, then add assets to the components... It was a nightmare....

I guess the next nightmare will be creating the GUI to access submeshes of ModelImports in the content browser so you can drag and drop them onto meshfilter and meshcomponent's in entities... Maybe it'll be a problem for another day! When I actually can be fucked...

Also separated the Create New Script Modal into the utils for the editor so that I can create a script from the content browser and the script component of an entity.