Prefabs, Instantiate in Scripting, Creating Assets, and More Physics!
- Louron Core -

Included Prefab to Asset Importer functions.

PREFAB INSTANTIATION AND SERIALISATION! You can now create prefabs by dragging an entity into the content browser and it will serialise the data into a ".lprefab" file type. I want to make a PrefabComponent that is attached to an entity when it is brought into the scene (ONLY IN EDITOR MODE) so that you can see if the prefab clone in scene has any changes made to it. When the scene is running, we don't actually need this. Maybe a better way would just to store changes made to entities in a map in the Louron Editor opposed to Louron Core? IDK more thinking on this one.

Scripts can instantiate prefabs aswell - you can create a "public Prefab" type in your script, and attach a prefab to it in the editor. Then if you call Instantiate in your script, and pass it that prefab, you can create it in scene!

Added Script callbacks for getting/setting parent objects, so you can child an entity to another entity, or you can detach children from parents, etc.

Moved NormalisePath to AssetManager - maybe I should just do a function like GetHandle(asset_type, file_path).

Added CreateAsset to AssetManager objects. This is for when I want to create a new asset from a filepath. This will be good for when I setup the filewatcher on the Content Browser.

Yet again fixed some issues with the physics system, copy/move was still not perfect for colliders and rigidbody. Also made sure that if the RigidbodyFlag_TransformUpdated is set in the PhysicsUpdate, that it skips updating the transform with the physics simulation for that frame. There was issues with Prefab instantiation where the rigidbody global pose was overwritting the manual transform updates when instantiating a prefab so most of them were instantiating at 0,0,0.

Added explicit copy/move functions for TransformComponent and HierarchyComponent for ENTT purposes to ensure that no data is being lost when it calls emplace_back and remove.

- Louron Editor -

Added a few new scripts, and new prefab files!

Added functionality to serialise and deserialise prefabs. Hierarchy Panel <--> Content Browser Panel.

Added functionality to attach prefabs from Content Browser Panel to Script Prefab Fields.

- Louron  Script Core -

Added functionality to Instantiate prefabs from scripts.

Added functionality to destroy entities.

Added functionality to get and set parent of Entity.

Made some adjustments to the math and vector classes, nothing major.