Point Light Shadow Mapping!
- Louron Core -

Implemented shadow mapping into Forward+ rendering pipeline. Currently can only support a limited number of shadow maps due to performance and memory concerns. Working on an algorithm locally to determine how to do something like cube map texture atlas for point lights, determining importance, how much res a point light cube map should get, etc.

Currently this works by creating an array of a fixed amount of cube maps. I will want to make this dynamic based on a global setting e.g., Shadow Quality, low, medium, high, etc. So maybe low settings only has 2 cube maps for point lights.

First I get the point lights which intersect with the frustum. Because I was already doing this for Forward+, this is free! I do alittle sorting to determine which light is the closest to the camera. Maybe I will have a cutoff worked into the global settings? Light position + radius distance to camera will be cutoff at a certain distance based on low medium or high settings?

Then I query the octree with the bounds of the point light to determine which meshes this point light intersects with. Note that a mesh outside the frustum could cast a shadow into the frustum, so the meshes need to be independent of the frustum, but dependent on being within the lights bounding sphere. Only meshes with the Cast Shadow option in their Mesh Renderer Component will be drawn in the shadow map.

Then the point light shadow FBO is bound and we loop through the shadow casting lights, calculate the projections of the 6 sides of a cube map, set the uniform values for the shader (courtesy to LearnOpenGL for this shader), then draw all sub_meshes that are in the lights shadow into the shadow cube map.

The pipeline also passes the cube map index to the shader in the UpdateSSBO stage of the pipeline so the shader knows which index in the cube map array to refer to.

Updated serialisation and deserialisation for mesh renderer components and point light components.

Fixed Octree query to not be a const function so I can actually query the octree using a BoundsAABB or a BoundsSphere.

Fixed issue with camera initialisation. When a scene is initialised, it will find the primary camera, and it will set the viewport which will force a recalculation of the projection and view matricies.

- Louron Editor -

Updated the one shader to rule them all! to implement shadows ofcourse.

I also updated the light attenuation function using Cem Yuksels proposed formula, and applied a scaling factor to the attenuation so its easier to control the intensity through the editor.

Implemented LearnOpenGL's shader for point light cube mapping.

Added a script called Orbit Dance, pretty funny for the goblin in the Shadow Test Scene.

Added functionality to turn the MeshRendererComponent on/off, and to make it cast shadows or not.

Fixed issue with editor camera receiving input when not hovering the scene viewport.

Fixed issue with OpenScene not actually loading the scene you select through the explorer dialog.